+++ R/common_functions.R
@@ -744,3 +744,201 @@ mean_comparisons_freq_anova = function(){
 plot_mean_comparisons_freq_anova = function(){
   
 }
+
+# map background for plot.data_agro, plot.data_network -----
+pmap = function(net, format, labels_on, labels_size){
+  # As it is not possible to use annotation_custom with polar coordinates (i.e. output from ggmap) in order to add pies on map,
+  # I decided to transfer ggmap output to a png that is inserted in a background of a plot with cartesian coordinates
+  # Note there is a change in the look of the map because of coordinates change ...
+  if( is_igraph(net) ){
+    d = ggnetwork(net, arrow.gap = 0)
+    
+    if( format == "bipart" ) {
+      d = d[which(d$type == "location"), c("lat", "long", "vertex.names")]
+      colnames(d)[ncol(d)] = "location"
+    } 
+    if( format == "unipart_location" ){
+      d = d[c("lat", "long", "vertex.names")]
+      colnames(d)[ncol(d)] = "location"
+    }
+  } else { d = net }
+
+  n = unique(d[, c("lat", "long", "location")]) 
+  n$lat = as.numeric(as.character(n$lat))
+  n$long = as.numeric(as.character(n$long))
+  n = na.omit(n)
+  center_location = c(mean(n$long), mean(n$lat))
+  map = get_map(location = center_location, source = "google", zoom = 6)
+  m = ggmap(map, extent = "device")
+  ggsave("tmp_map.png", m, width = 1, height = 1) # get a perfect square
+  p = ggplot(mtcars, aes(wt, mpg)) + geom_point(size = -10) # support for the map background
+  p = p + coord_cartesian(xlim = range(m$data$lon), ylim = range(m$data$lat), expand = FALSE)
+  img = readPNG("tmp_map.png")
+  pmap = p + annotation_custom(rasterGrob(img, width = unit(1,"npc"), height = unit(1,"npc")), 
+                               -Inf, Inf, -Inf, Inf) # change in the look of the map because of coordinates changes
+  pmap = pmap + xlab("long") + ylab("lat")
+  file.remove("tmp_map.png")
+  if( !is.null(labels_on) ){
+    if( labels_on == "location" ){ 
+      pmap = pmap + geom_nodelabel_repel(data = n, aes(x = long, y = lat, label = location), size = labels_size, inherit.aes = FALSE) 
+    }
+  }
+  return(pmap)
+}
+
+
+# Add_pies for plot.data_agro, plot.data_network -----
+add_pies = function(p, n, format, plot_type, data_to_pie, variable, pie_size){
+  # script adapted from 
+  # Pies On A Map, Demonstration script, By QDR : 
+  #   https://qdrsite.wordpress.com/2016/06/26/pies-on-a-map/
+  # Guangchuang YU code :
+  #   https://cran.r-project.org/web/packages/ggimage/vignettes/ggimage.html#geom_subview
+  #   https://github.com/GuangchuangYu/ggimage/blob/master/R/geom_subview.R
+  
+  # p : network or map
+  # n : network object from igraph or data frame with coordinates
+  # data_to_pie : data with the variables
+  
+  # add a invisible point with variable value to get the legend of pies + set the legend
+  colnames(data_to_pie)[which(colnames(data_to_pie) == variable)] = "variable"
+  
+  col_low = "red" # "#132B43"
+  col_high = "green" # "#56B1F7"
+  
+  if( is.numeric(data_to_pie$variable) ) {  
+    p = p + geom_point(data = data_to_pie, x = 0, y = 0, size = -10, aes(fill = variable), inherit.aes = FALSE)
+    p = p + scale_fill_continuous(low = col_low, high = col_high)
+    scale_ok = scales::seq_gradient_pal(low = col_low, high = col_high)(seq(0, 1, length.out = nrow(data_to_pie)))
+    s = seq(min(data_to_pie$variable, na.rm = TRUE), max(data_to_pie$variable, na.rm = TRUE), length.out = nrow(data_to_pie))
+    data_to_pie$scale_col = sapply(data_to_pie$variable, function(x){scale_ok[which(s >= x)[1]]})
+  }
+  
+  if( is.factor(data_to_pie$variable) ) {  
+    p = p + geom_point(data = data_to_pie, x = -10, y = 10, aes(shape = variable, fill = variable), inherit.aes = FALSE) 
+    p = p + scale_shape_manual(values = rep(22, nlevels(data_to_pie$variable)))
+    scale_ok = scales::seq_gradient_pal(low = col_low, high = col_high)(seq(0, 1, length.out = nlevels(data_to_pie$variable)))
+    p = p + scale_fill_manual(values = scale_ok)
+    s = seq(1, nlevels(data_to_pie$variable))
+    data_to_pie$scale_col = sapply(as.numeric(data_to_pie$variable), function(x){scale_ok[which(s >= x)[1]]})
+  }
+  
+  
+  # Set colnames for next step according to plot type and get range for x and y
+  if( plot_type == "map" ) { 
+    colnames(data_to_pie)[which(colnames(data_to_pie) == "location")] = "id_ok" 
+    xmin = min(p$coordinates$limits$x); xmax = max(p$coordinates$limits$x)
+    ymin = min(p$coordinates$limits$y); ymax = max(p$coordinates$limits$y)
+  }
+  
+  if( plot_type == "network" ){ 
+    colnames(data_to_pie)[which(colnames(data_to_pie) == "id")] = "id_ok" 
+    xmin = min(p$data$x); xmax = max(p$data$x)
+    ymin = min(p$data$y); ymax = max(p$data$y)
+  }
+  
+  # Create a list of ggplot objects. Each one is the pie chart for each site with all labels removed.
+  pies <- dlply(data_to_pie, .(id_ok), function(z){
+    z = arrange(z, variable)
+    s_col = z$scale_col; names(s_col) = z$variable
+    s_col = s_col[unique(names(s_col))]
+    ggplot(z, aes(x = factor(1), fill = factor(variable))) +
+      geom_bar(width = 1) +
+      coord_polar(theta = "y") +
+      scale_fill_manual(values = s_col) +
+      theme(axis.line=element_blank(),
+            axis.text.x=element_blank(),
+            axis.text.y=element_blank(),
+            axis.ticks=element_blank(),
+            axis.title.x=element_blank(),
+            axis.title.y=element_blank(),
+            legend.position="none",
+            panel.background=element_blank(),
+            panel.border=element_blank(),
+            panel.grid.major=element_blank(),
+            panel.grid.minor=element_blank(),
+            plot.background=element_blank()) 
+  }
+  )
+  
+  # Get coordinates of each pie and select pies
+  if( is_igraph(n) ){
+    d = ggnetwork(n, arrow.gap = 0)
+  } else { d = n }
+  v_id = c(unique(as.character(data_to_pie$id_ok)))
+  
+  if( plot_type == "map" & format == "bipart" ){
+    d = droplevels(d[which(d$type == "location"),])
+    v_ok = v_id[which(is.element(v_id, as.character(d$vertex.names) ))]
+    v_not_ok = v_id[which(!is.element(v_id, as.character(d$vertex.names) ))]
+  } 
+  if( plot_type == "map" & format == "unipart_location" ){
+    v_ok = v_id[which(is.element(v_id, as.character(d$vertex.names) ))]
+    v_not_ok = v_id[which(!is.element(v_id, as.character(d$vertex.names) ))]
+  } 
+  if( plot_type == "map" & format == "unipart_sl" ){
+    v_ok = v_id[which(is.element(v_id, as.character(d$location) ))]
+    v_not_ok = v_id[which(!is.element(v_id, as.character(d$location) ))]
+  }    
+  if( plot_type == "network" & format == "unipart_sl" ){
+    v_ok = v_id[which(is.element(v_id, as.character(d$vertex.names) ))]
+    v_not_ok = v_id[which(!is.element(v_id, as.character(d$vertex.names) ))]
+  } 
+  if( plot_type == "map" & format == "data_agro" ){
+    v_ok = v_id[which(is.element(v_id, as.character(d$location) ))]
+    v_not_ok = v_id[which(!is.element(v_id, as.character(d$location) ))]
+  }    
+  pies = pies[v_ok]
+  
+  if( length(v_ok) == 0) { 
+    warning("In the data with the variable, no id exist in the data with coordinates and therefore no pies are displayed") 
+  }
+  if( length(v_ok) < length(v_id) ){ 
+    warning("In the data with the variable, the following id does not exist in the data with the coordinates: ", paste(v_not_ok, collapse = ", ")) 
+  }
+  
+  if( plot_type == "map" ){ 
+    if( length(v_ok) > 0 ) {
+      if( format == "bipart" ) {
+        d = ggnetwork(n, arrow.gap = 0)
+        d = d[which(d$type == "location"), c("lat", "long", "vertex.names")]
+        colnames(d)[ncol(d)] = "location"
+      } 
+      if( format == "unipart_location" ){
+        d = ggnetwork(n, arrow.gap = 0)
+        d = d[c("lat", "long", "vertex.names")]
+        colnames(d)[ncol(d)] = "location"
+      }
+      d = unique(d[, c("lat", "long", "location")])
+      
+      piecoords = lapply(names(pies), function(x){
+        c(x = as.numeric(as.character(d[which(d$location == x), "long"])), 
+          y = as.numeric(as.character(d[which(d$location == x), "lat"]))
+        )
+      } 
+      )
+    }
+  }
+  
+  
+  if( plot_type == "network" ){ 
+    if( length(v_ok) > 0 ) {
+      piecoords = lapply(names(pies), function(x){
+        c(x = unique(p$data[which(p$data$vertex.names == x), "x"]), y = unique(p$data[which(p$data$vertex.names == x), "y"]))
+      }
+      )
+    }
+  }
+  
+  # add pies on plot
+  if( length(v_ok) > 0 ) {
+    for(i in 1:length(pies)){
+      p = p + geom_subview(x = piecoords[[i]]["x"], y = piecoords[[i]]["y"], 
+                           subview = pies[[i]], 
+                           width = (xmax-xmin)*pie_size, height = (ymax-ymin)*pie_size)
+    }
+  }
+  
+  return(p)
+}
+